+++ src/main/java/racingcar/car/CarManager.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import racingcar.car.Car;
 import racingcar.constant.Number;
 import racingcar.iomanagement.InputScanner;
 import racingcar.iomanagement.Validator;
